Exceptions shouldn't be objects, redundant toType calls, redundant qualifiers, redundant semicolons

+++ src/main/kotlin/net/ccbluex/liquidbounce/deeplearn/data/TrainingData.kt
@@ -71,7 +71,7 @@ data class TrainingData(
             targetDiff.horizontalLength().toFloat() + playerDiff.horizontalLength().toFloat(),
 
             // Distance
-            distance.toFloat()
+            distance
         )
 
     val asOutput
